+++ packages/cli/test/util/load-plan.test.ts
@@ -4,11 +4,7 @@ import { createMockLogger } from '@openfn/logger';
 import type { Job } from '@openfn/lexicon';
 
 import loadPlan from '../../src/util/load-plan';
-import {
-  collectionsEndpoint,
-  collectionsVersion,
-  Opts,
-} from '../../src/options';
+import { Opts } from '../../src/options';
 
 const logger = createMockLogger(undefined, { level: 'debug' });
 
@@ -38,6 +34,7 @@ test.beforeEach(() => {
       jobs: [{ id: 'a', expression: 'x()' }],
     }),
     'test/wf.json': JSON.stringify(sampleXPlan),
+    'test/wf-flat.json': JSON.stringify(sampleXPlan.workflow),
     'test/wf-err.json': '!!!',
   });
 });
@@ -163,7 +160,7 @@ test.serial(
   }
 );
 
-test.serial('xplan: load a plan from workflow path', async (t) => {
+test.serial('xplan: load an old-style plan from workflow path', async (t) => {
   const opts = {
     workflowPath: 'test/wf.json',
     expandAdaptors: true,
@@ -174,6 +171,22 @@ test.serial('xplan: load a plan from workflow path', async (t) => {
 
   t.truthy(plan);
   t.deepEqual(plan, sampleXPlan);
+});
+
+test.serial('xplan: load a new flat plan from workflow path', async (t) => {
+  const opts = {
+    workflowPath: 'test/wf-flat.json',
+    expandAdaptors: true,
+    plan: {},
+  };
+
+  const plan = await loadPlan(opts, logger);
+
+  t.truthy(plan);
+  t.deepEqual(plan, {
+    options: {}, // no options here!
+    workflow: sampleXPlan.workflow,
+  });
 });
 
 test.serial('xplan: expand adaptors', async (t) => {
@@ -429,3 +442,217 @@ test.serial('xplan: append collections', async (t) => {
     collections_token: opts.apiKey,
   });
 });
+
+test.serial(
+  'xplan: append collections to existing credential object',
+  async (t) => {
+    const opts = {
+      workflowPath: 'test/wf.json',
+      collectionsVersion: '1.1.1',
+      collectionsEndpoint: 'https://localhost:4000/',
+      apiKey: 'abc',
+    };
+
+    const plan = createPlan([
+      {
+        id: 'a',
+        expression: 'collections.get()',
+        adaptors: ['@openfn/language-common@1.0.0'],
+        configuration: {
+          x: 1,
+        },
+      },
+    ]);
+
+    mock({
+      'test/wf.json': JSON.stringify(plan),
+    });
+
+    const result = await loadPlan(opts, logger);
+    t.truthy(result);
+
+    const step = result.workflow.steps[0] as Job;
+    t.deepEqual(step.adaptors, [
+      '@openfn/language-common@1.0.0',
+      '@openfn/language-collections@1.1.1',
+    ]);
+
+    t.deepEqual(step.configuration, {
+      collections_endpoint: `${opts.collectionsEndpoint}/collections`,
+      collections_token: opts.apiKey,
+      x: 1,
+    });
+  }
+);
+
+test.serial(
+  'xplan: load a workflow.yaml without top workflow key',
+  async (t) => {
+    mock({
+      'test/wf.yaml': `
+name: wf
+steps:
+  - id: a
+    adaptors: []
+    expression: x()
+`,
+    });
+    const opts = {
+      path: 'test/wf.yaml',
+    };
+
+    const plan = await loadPlan(opts, logger);
+
+    t.truthy(plan);
+    // Note that options are lost in this design!
+    t.deepEqual(plan, { workflow: sampleXPlan.workflow, options: {} });
+  }
+);
+
+test.serial(
+  'xplan: load a workflow.yaml without top workflow key and options',
+  async (t) => {
+    mock({
+      'test/wf.yaml': `
+name: wf
+steps:
+  - id: a
+    adaptors: []
+    expression: x()
+options:
+  start: x
+`,
+    });
+    const opts = {
+      path: 'test/wf.yaml',
+    };
+
+    const plan = await loadPlan(opts, logger);
+
+    t.truthy(plan);
+    // Note that options are lost in this design!
+    t.deepEqual(plan, {
+      workflow: sampleXPlan.workflow,
+      options: { start: 'x' },
+    });
+  }
+);
+
+test.serial('xplan: load a workflow.yaml with top workflow key', async (t) => {
+  mock({
+    'test/wf.yaml': `
+workflow:
+  name: wf
+  steps:
+    - id: a
+      adaptors: []
+      expression: x()
+options:
+  start: a
+`,
+  });
+  const opts = {
+    path: 'test/wf.yaml',
+  };
+
+  const plan = await loadPlan(opts, logger);
+
+  t.truthy(plan);
+  t.deepEqual(plan, sampleXPlan);
+});
+
+test.serial('xplan: load a workflow through a Workspace', async (t) => {
+  mock({
+    '/tmp/workflows/wf.yaml': `
+id: wf
+steps:
+  - id: a
+    expression: x()
+`,
+    '/tmp/openfn.yaml': `
+dirs:
+  workflows: /tmp/workflows
+`,
+  });
+
+  const opts = {
+    // TODO is worked out through yargs via the inputPath option
+    workflowName: 'wf',
+    workspace: '/tmp',
+  };
+
+  const plan = await loadPlan(opts, logger);
+  t.truthy(plan);
+  t.deepEqual(plan, {
+    workflow: {
+      id: 'wf',
+      steps: [{ id: 'a', expression: 'x()', adaptors: [] }],
+      history: [],
+    },
+    options: {},
+  });
+});
+
+test.serial('xplan: throw if a named workflow does not exist', async (t) => {
+  mock({
+    '/tmp/workflows/wf.yaml': `
+id: wf
+steps:
+  - id: a
+    expression: x()
+`,
+    '/tmp/openfn.yaml': `
+dirs:
+  workflows: /tmp/workflows
+`,
+  });
+
+  const opts = {
+    workflowName: 'JAM',
+    workspace: '/tmp',
+  };
+
+  await t.throwsAsync(() => loadPlan(opts, logger), {
+    message: /could not find workflow "jam"/i,
+  });
+});
+
+test.serial(
+  'xplan: load a workflow through a project .yaml and apply the credentials map by default',
+  async (t) => {
+    mock({
+      '/tmp/workflows/wf.yaml': `
+id: wf
+steps:
+  - id: a
+    expression: x()
+start: a
+`,
+      '/tmp/openfn.yaml': `
+credentials: /creds.yaml
+dirs:
+  workflows: /tmp/workflows
+`,
+      '/creds.yaml': `x: y`,
+    });
+    const opts = {
+      workflowName: 'wf',
+      workspace: '/tmp',
+    };
+
+    const plan = await loadPlan(opts, logger);
+
+    t.truthy(plan);
+    t.deepEqual(plan, {
+      workflow: {
+        id: 'wf',
+        steps: [{ id: 'a', expression: 'x()', adaptors: [] }],
+        history: [],
+        start: 'a',
+      },
+      options: {},
+    });
+
+    t.is(opts.credentials, '/creds.yaml');
+  }
+);