+++ packages/project/src/parse/from-project.ts
@@ -3,7 +3,7 @@ import * as l from '@openfn/lexicon';
 import Project from '../Project';
 import ensureJson from '../util/ensure-json';
 import { Provisioner } from '@openfn/lexicon/lightning';
-import fromAppState from './from-app-state';
+import fromAppState, { fromAppStateConfig } from './from-app-state';
 import { WithMeta } from '../Workflow';
 
 // Load a project from any JSON or yaml representation
@@ -32,25 +32,23 @@ export default (
   // first ensure the data is in JSON format
   let rawJson = ensureJson<any>(data);
 
-  let json;
   if (rawJson.cli?.version ?? rawJson.version /*deprecated*/) {
     // If there's any version key at all, its at least v2
-    json = from_v2(rawJson as SerializedProject);
-  } else {
-    json = from_v1(rawJson as Provisioner.Project);
+    return new Project(from_v2(rawJson as SerializedProject), config);
   }
 
-  return new Project(json, config);
+  return from_v1(rawJson as Provisioner.Project, config as fromAppStateConfig);
 };
 
-const from_v1 = (data: Provisioner.Project) => {
-  // TODO is there any way to look up the config file?
-  // But we have no notion of a working dir here
-  // Maybe there are optional options that can be provided
-  // by from fs or from path
-  return fromAppState(data);
+// TODO test that config (alias) works
+const from_v1 = (
+  data: Provisioner.Project,
+  config: fromAppStateConfig = {}
+) => {
+  return fromAppState(data, {}, config);
 };
 
+// TODO this should return a Project really!
 const from_v2 = (data: SerializedProject) => {
   // nothing to do
   // (When we add v3, we'll ned to migrate through this)