+++ packages/lightning-mock/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/lightning-mock
 
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies [b262d10]
+- Updated dependencies [147a431]
+  - @openfn/runtime@1.8.1
+  - @openfn/engine-multi@1.10.1
+
 ## 2.4.1
 
 ### Patch Changes